+++ html/mod_menu/default_separator.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * @package     Joomla.Site
+ * @subpackage  mod_menu
+ *
+ * @copyright   (C) 2009 Open Source Matters, Inc. <https://www.joomla.org>
+ * @license     GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+defined('_JEXEC') or die;
+
+use Joomla\CMS\HTML\HTMLHelper;
+
+$title      = $item->anchor_title ? ' title="' . $item->anchor_title . '"' : '';
+$anchor_css = $item->anchor_css ?: '';
+$linktype   = $item->title;
+
+if ($item->menu_image)
+{
+	if ($item->menu_image_css)
+	{
+		$image_attributes['class'] = $item->menu_image_css;
+		$linktype = HTMLHelper::_('image', $item->menu_image, $item->title, $image_attributes);
+	}
+	else
+	{
+		$linktype = HTMLHelper::_('image', $item->menu_image, $item->title);
+	}
+
+	if ($itemParams->get('menu_text', 1))
+	{
+		$linktype .= '<span class="image-title">' . $item->title . '</span>';
+	}
+}
+?>
+
+<?php if ($showAll && $item->deeper) : ?>
+	<a href="#" class="mod-menu__separator separator <?php echo $anchor_css; ?>"<?php echo $title; ?>><?php echo $linktype; ?></a>
+<?php else : ?>
+	<span class="mod-menu__separator separator <?php echo $anchor_css; ?>"<?php echo $title; ?>><?php echo $linktype; ?></span>
+<?php endif; ?>